(refractor) hooks tests
